+++ bevy_asset_loader/src/asset_loader/dynamic_asset_systems.rs
@@ -7,6 +7,7 @@ use bevy::ecs::system::{Res, SystemState};
 use bevy::ecs::world::World;
 
 pub(crate) fn load_dynamic_asset_collections<S: StateData>(world: &mut World) {
+    #[allow(clippy::type_complexity)]
     let mut system_state: SystemState<(
         ResMut<DynamicAssetCollections<S>>,
         ResMut<State<LoadingState>>,
@@ -40,6 +41,7 @@ pub(crate) fn load_dynamic_asset_collections<S: StateData>(world: &mut World) {
 }
 
 pub(crate) fn check_dynamic_asset_collections<S: StateData>(world: &mut World) {
+    #[allow(clippy::type_complexity)]
     let mut system_state: SystemState<(
         Res<AssetServer>,
         ResMut<LoadingAssetHandles<S>>,