+++ bevy_asset_loader/tests/ui_with_2d_3d_features/key_and_any_other_attribute.rs
@@ -33,6 +33,6 @@ struct OptionalDynamic {
 // dynamic folder
 #[derive(AssetCollection)]
 struct FolderAndKey {
-    #[asset(folder, key = "test")]
+    #[asset(collection, key = "test")]
     test: Vec<HandleUntyped>,
 }